Make custom sprite work. Refactor sprite and entity to accomodate new sprite class. Implement base for particle system.

+++ src/pelmeni/graphics/AnimationState.hpp
@@ -16,7 +16,7 @@ namespace p2d { namespace graphics {
 
         bool update(const sf::Time& time);
         Frame& getCurrentFrame() const;
-        sf::Rect<int>& getCurrentFrameRect() const;
+        sf::Rect<float>& getCurrentFrameRect() const;
         sf::Time& getCurrentFrameDuration() const;
     private:
         size_t currentFrameIndex;
